refactor: recreate SearchInput as component
