+++ src/pages/home/components/HomeSearch/HomeSearch.styles.ts
@@ -35,12 +35,3 @@ export const DescriptionContent = styled.div`
     font-size: 28px;
   }
 `;
-
-export const SearchInput = styled.input`
-  width: 80%;
-  font-size: 16px;
-  padding: 16px 20px;
-  margin-bottom: 10px;
-  border-radius: 50px;
-  border: 2px solid ${({ theme }) => theme.colors.purple} !important;
-`;
